requirements.txt
Django==1.4
Fabric==1.4.1
South==0.7.4
psycopg2==2.4.5
pycrypto==2.5
pytz==2012c
ssh==1.7.13
wsgiref==0.1.2
django-crispy-forms==1.1.3
#django-userena==1.1.0
-e git+https://github.com/zerok/django-userena@e6b518632a04f1fd3343885c597a07c75dbb8953#egg=django-userena
django-recaptcha==0.0.4
python-postmark==0.3.1
gunicorn==0.14.3
django-debug-toolbar==0.9.4
Markdown==2.1.1
markdown2==1.4.2
PIL==1.1.7
raven==1.9.1